@@ -1,6 +1,6 @@
 <?php
 /*
- *  Last Modified: 10/04/23, 11:50 PM
+ *  Last Modified: 12/06/23, 11:50 PM
  *  Copyright (c) 2023
  *  -created by Ariful Islam
  *  -All Rights Preserved By
@@ -16,15 +16,15 @@ use Xenon\LaravelBDSms\Request;
 use Xenon\LaravelBDSms\Sender;
 
 /**
- * ====================================================================================================================
- * Alpha SMS is an advanced SMS service provider in Bangladesh with multiple services such as Bulk SMS, OTP Notification,
+ * =====================================================================================================================================
+ * SMS.NET.BD( Alpha ) SMS is an advanced SMS service provider in Bangladesh with multiple services such as Bulk SMS, OTP Notification,
  * SMS Gateway, Marketing, Masking SMS, etc. It provides a powerful and easy-to-use SMS API for developers that you can use
  * to integration with Software Applications & Websites.
- * =========================================================================================================================
+ * ======================================================================================================================================
  */
-class Alpha extends AbstractProvider
+class SmsNetBD extends AbstractProvider
 {
-    private $apiEndpoint ='https://api.sms.net.bd/sendsms';
+    private string $apiEndpoint = 'https://api.sms.net.bd/sendsms';
 
     /**
      * Alpha SMS constructor.
@@ -69,7 +69,7 @@ class Alpha extends AbstractProvider
             $query['sender_id'] = $config['sender_id'];
         }
         if (is_array($mobile)) {
-            $query['to'] = implode(',', $mobile);
+            $query['to'] =  implode(',', $mobile);
         }
 
         $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
