@@ -30,7 +30,7 @@ class RenderException extends \Exception
      * Render the exception into an HTTP response.
      *
      * @param Request $request
-     * @return Response
+     * @return void
      */
     public function render(Request $request)
     {
