@@ -38,12 +38,25 @@ class Helper
      * @since v1.0.12
      * @version v1.0.43.1-dev
      */
-    public static function checkMobileNumberPrefixExistence($mobile)
+    public static function checkMobileNumberPrefixExistence(string $mobile)
     {
         $prefix = substr($mobile, 0, 3);
         if ($prefix === '880') {
             return $mobile;
         }
         return '88' . $mobile;
+    }
+
+    /**
+     * @param string $text
+     * @return string
+     * @since v1.0.52.0-beta
+     * @version v1.0.52.0-beta
+     */
+    public static function ensureNumberStartsWith88(string $text): string {
+        if (!str_starts_with($text, '88')) {
+            $text = '88' . $text;
+        }
+        return $text;
     }
 }
