@@ -1,27 +1,25 @@
 <?php
 /*
- *  Last Modified: 6/29/21, 12:06 AM
- *  Copyright (c) 2021
+ *  Last Modified: 02/02/23, 11:50 PM
+ *  Copyright (c) 2023
  *  -created by Ariful Islam
  *  -All Rights Preserved By
  *  -If you have any query then knock me at
  *  arif98741@gmail.com
- *  See my profile @ https://github.com/arif98741
+ *  See my profile @https://github.com/arif98741
  */
 
 namespace Xenon\LaravelBDSms\Provider;
 
 use GuzzleHttp\Exception\GuzzleException;
-use Xenon\LaravelBDSms\Handler\ParameterException;
 use Xenon\LaravelBDSms\Handler\RenderException;
 use Xenon\LaravelBDSms\Request;
 use Xenon\LaravelBDSms\Sender;
 
-class BulkSmsBD extends AbstractProvider
+class CustomGateway extends AbstractProvider
 {
-    private string $apiEndpoint = 'https://bulksmsbd.net/api/smsapi';
     /**
-     * BulkSmsBD constructor.
+     * Custom Gateway constructor.
      * @param Sender $sender
      */
     public function __construct(Sender $sender)
@@ -31,52 +29,46 @@ class BulkSmsBD extends AbstractProvider
 
     /**
      * Send Request To Api and Send Message
-     * @throws GuzzleException|RenderException
+     * @throws RenderException|GuzzleException
      */
     public function sendRequest()
     {
-        $number = $this->senderObject->getMobile();
+        $mobile = $this->senderObject->getMobile();
         $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
         $queueName = $this->senderObject->getQueueName();
         $tries=$this->senderObject->getTries();
         $backoff=$this->senderObject->getBackoff();
+        $query = $config;
 
-        $query = [
-            'api_key' => $config['api_key'],
-            'senderid' => $config['senderid'],
-            'type' => 'text',
-            'number' => $number,
-            'message' => $text,
-        ];
+        $requestObject = new Request($this->senderObject->url, $query, $queue, [], $queueName,$tries,$backoff);
 
-        if (array_key_exists('senderid', $config)) {
-            $query ['senderid'] = $config['senderid'];
+        if (isset($this->senderObject->headers)) {
+            $requestObject->setHeaders($this->senderObject->headers);
+            $this->senderObject->contentTypeJson && $requestObject->setContentTypeJson(true);
         }
 
-        $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
-        $response = $requestObject->get();
+        $response = $this->senderObject->method === 'post' ? $requestObject->post() : $requestObject->get();
+
         if ($queue) {
             return true;
         }
 
         $body = $response->getBody();
         $smsResult = $body->getContents();
-
-        $data['number'] = $number;
+        $data['number'] = $mobile;
         $data['message'] = $text;
         return $this->generateReport($smsResult, $data)->getContent();
     }
 
     /**
-     * @throws ParameterException
+     * @throws RenderException
      */
-    public function errorException()
+    final public function errorException(): void
     {
-        if (!array_key_exists('api_key', $this->senderObject->getConfig())) {
-            throw new ParameterException('api_key key is absent in configuration');
+        if (!isset($this->senderObject->url)) {
+            throw new RenderException("Url missing for custom gateway. Use setUrl() to set sms gateway endpoint ");
         }
-
     }
 }
