@@ -17,6 +17,8 @@ use Xenon\LaravelBDSms\Sender;
 
 class DianaHost extends AbstractProvider
 {
+    private string $apiEndpoint ='http://esms.dianahost.com/smsapi';
+
     /**
      * DianaHost constructor.
      * @param Sender $sender
@@ -35,6 +37,9 @@ class DianaHost extends AbstractProvider
         $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
+        $queueName = $this->senderObject->getQueueName();
+        $tries=$this->senderObject->getTries();
+        $backoff=$this->senderObject->getBackoff();
 
         $query = [
             'api_key' => $config['api_key'],
@@ -45,7 +50,7 @@ class DianaHost extends AbstractProvider
         ];
 
 
-        $requestObject = new Request('http://esms.dianahost.com/smsapi', $query, $queue);
+        $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
         $response = $requestObject->get();
         if ($queue) {
             return true;
