@@ -39,6 +39,9 @@ class DianaSms extends AbstractProvider
         $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
+        $queueName = $this->senderObject->getQueueName();
+        $tries=$this->senderObject->getTries();
+        $backoff=$this->senderObject->getBackoff();
 
         $query = [
             'ApiKey' => $config['ApiKey'],
@@ -48,7 +51,7 @@ class DianaSms extends AbstractProvider
             'Message' => $text,
         ];
 
-        $requestObject = new Request('https://q.dianasms.com/api/v2/SendSMS', $query, $queue);
+        $requestObject = new Request('https://q.dianasms.com/api/v2/SendSMS', $query, $queue, [], $queueName,$tries,$backoff);
         $response = $requestObject->get();
         if ($queue) {
             return true;
