@@ -1,7 +1,7 @@
 <?php
 /*
- *  Last Modified: 6/29/21, 12:06 AM
- *  Copyright (c) 2021
+ *  Last Modified: 6/16/22, 12:56 AM
+ *  Copyright (c) 2022
  *  -created by Ariful Islam
  *  -All Rights Preserved By
  *  -If you have any query then knock me at
@@ -11,16 +11,18 @@
 
 namespace Xenon\LaravelBDSms\Provider;
 
+use GuzzleHttp\Exception\GuzzleException;
+use Xenon\LaravelBDSms\Handler\ParameterException;
 use Xenon\LaravelBDSms\Handler\RenderException;
 use Xenon\LaravelBDSms\Request;
 use Xenon\LaravelBDSms\Sender;
 
-class MDL extends AbstractProvider
+class DurjoySoft extends AbstractProvider
 {
-    private string $apiEndpoint  = 'http://premium.mdlsms.com/smsapi';
+    private string $apiEndpoint = 'https://smsp.durjoysoft.com/api/sms';
 
     /**
-     * MDL constructor.
+     * DurjoySoft constructor.
      * @param Sender $sender
      */
     public function __construct(Sender $sender)
@@ -30,6 +32,9 @@ class MDL extends AbstractProvider
 
     /**
      * Send Request To Api and Send Message
+     * @return bool|string
+     * @throws GuzzleException
+     * @throws RenderException
      */
     public function sendRequest()
     {
@@ -42,11 +47,11 @@ class MDL extends AbstractProvider
         $backoff=$this->senderObject->getBackoff();
 
         $query = [
-            'api_key' => $config['api_key'],
-            'type' => $config['type'],
-            'senderid' => $config['senderid'],
-            'contacts' => $number,
-            'msg' => $text,
+            'ApiKey' => $config['ApiKey'],
+            'SenderID' => $config['SenderID'],
+            'number' => $number,
+            'sms' => $text,
+            'IsUnicode' => 2,
         ];
 
         $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
@@ -64,19 +69,16 @@ class MDL extends AbstractProvider
     }
 
     /**
-     * @throws RenderException
+     * @throws ParameterException
      */
     public function errorException()
     {
-        if (!array_key_exists('api_key', $this->senderObject->getConfig())) {
-            throw new RenderException('api_key is absent in configuration');
+        if (!array_key_exists('ApiKey', $this->senderObject->getConfig())) {
+            throw new ParameterException('ApiKey is absent in configuration');
         }
-        if (!array_key_exists('type', $this->senderObject->getConfig())) {
-            throw new RenderException('type key is absent in configuration');
+        if (!array_key_exists('SenderID', $this->senderObject->getConfig())) {
+            throw new ParameterException('SenderID key is absent in configuration');
         }
-        if (!array_key_exists('senderid', $this->senderObject->getConfig())) {
-            throw new RenderException('senderid key is absent in configuration');
-        }
-
     }
+
 }
