@@ -1,7 +1,7 @@
 <?php
 /*
- *  Last Modified: 6/29/21, 12:06 AM
- *  Copyright (c) 2021
+ *  Last Modified: 10/04/23, 11:50 PM
+ *  Copyright (c) 2023
  *  -created by Ariful Islam
  *  -All Rights Preserved By
  *  -If you have any query then knock me at
@@ -11,18 +11,16 @@
 
 namespace Xenon\LaravelBDSms\Provider;
 
-use GuzzleHttp\Exception\GuzzleException;
-use Xenon\LaravelBDSms\Handler\ParameterException;
 use Xenon\LaravelBDSms\Handler\RenderException;
 use Xenon\LaravelBDSms\Request;
 use Xenon\LaravelBDSms\Sender;
 
-class Mobireach extends AbstractProvider
+class Muthofun extends AbstractProvider
 {
-    private string $apiEndpoint = 'https://api.mobireach.com.bd/SendTextMessage';
+    private string $apiEndpoint = 'https://sysadmin.muthobarta.com/api/v1/send-sms';
 
     /**
-     * Mobireach constructor.
+     * Muthofun constructor.
      * @param Sender $sender
      */
     public function __construct(Sender $sender)
@@ -32,13 +30,11 @@ class Mobireach extends AbstractProvider
 
     /**
      * Send Request To Api and Send Message
-     * @return bool|string
-     * @throws GuzzleException
      * @throws RenderException
      */
     public function sendRequest()
     {
-        $number = $this->senderObject->getMobile();
+        $mobile = $this->senderObject->getMobile();
         $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
@@ -47,44 +43,45 @@ class Mobireach extends AbstractProvider
         $backoff = $this->senderObject->getBackoff();
 
         $query = [
-            'Username' => $config['Username'],
-            'Password' => $config['Password'],
-            'From' => $config['From'],
-            'To' => $number,
-            'Message' => $text,
+            'sender_id' => $config['sender_id'],
+            'remove_duplicate' => true,
+            'receiver' => $mobile,
+            'message' => $text,
         ];
 
+        if (is_array($mobile)) {
+            $query['receiver'] = implode(',', $mobile);
+        }
+
         $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName, $tries, $backoff);
-        $response = $requestObject->get();
+
+        if (!str_starts_with($config['api_key'], "Token ")) {
+            $config['api_key'] = "Token " . $config['api_key'];
+        }
+        $requestObject->setHeaders(['Authorization' => $config['api_key']])->setContentTypeJson(true);
+
+        $response = $requestObject->post();
         if ($queue) {
             return true;
         }
-
         $body = $response->getBody();
         $smsResult = $body->getContents();
-
-        $data['number'] = $number;
+        $data['number'] = $mobile;
         $data['message'] = $text;
         return $this->generateReport($smsResult, $data)->getContent();
     }
 
     /**
-     * @throws ParameterException
+     * @throws RenderException
      */
-    public function errorException()
+    public function errorException(): void
     {
-        if (!array_key_exists('Username', $this->senderObject->getConfig())) {
-            throw new ParameterException('Username is absent in configuration');
+        if (!array_key_exists('sender_id', $this->senderObject->getConfig())) {
+            throw new RenderException('sender_id key is absent in configuration');
         }
-
-        if (!array_key_exists('Password', $this->senderObject->getConfig())) {
-            throw new ParameterException('Password is absent in configuration');
-        }
-
-        if (!array_key_exists('From', $this->senderObject->getConfig())) {
-            throw new ParameterException('From is absent in configuration');
+        if (!array_key_exists('api_key', $this->senderObject->getConfig())) {
+            throw new RenderException('api_key key is absent in configuration');
         }
 
     }
-
 }
