@@ -11,12 +11,16 @@
 
 namespace Xenon\LaravelBDSms\Provider;
 
+use GuzzleHttp\Exception\GuzzleException;
 use Xenon\LaravelBDSms\Handler\ParameterException;
+use Xenon\LaravelBDSms\Handler\RenderException;
 use Xenon\LaravelBDSms\Request;
 use Xenon\LaravelBDSms\Sender;
 
 class NovocomBd extends AbstractProvider
 {
+    private string $apiEndpoint = 'https://sms.novocom-bd.com/api/v2/SendSMS';
+
     /**
      * Novocom constructor.
      * @param Sender $sender
@@ -28,9 +32,9 @@ class NovocomBd extends AbstractProvider
 
     /**
      * Send Request To Api and Send Message
-     * @return bool|mixed|string
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Xenon\LaravelBDSms\Handler\RenderException
+     * @return bool|string
+     * @throws GuzzleException
+     * @throws RenderException
      */
     public function sendRequest()
     {
@@ -38,6 +42,9 @@ class NovocomBd extends AbstractProvider
         $number = $this->senderObject->getMobile();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
+        $queueName = $this->senderObject->getQueueName();
+        $tries=$this->senderObject->getTries();
+        $backoff=$this->senderObject->getBackoff();
 
         $query = [
             'ApiKey' => $config['ApiKey'],
@@ -48,7 +55,7 @@ class NovocomBd extends AbstractProvider
             'Is_Unicode' => true,
         ];
 
-        $requestObject = new Request('https://sms.novocom-bd.com/api/v2/SendSMS', $query, $queue);
+        $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
         $response = $requestObject->get();
         if ($queue) {
             return true;
