@@ -19,6 +19,8 @@ use Xenon\LaravelBDSms\Sender;
 
 class SmartLabSms extends AbstractProvider
 {
+    private string $apiEndpoint = 'https://labapi.smartlabsms.com/smsapi';
+
     /**
      * SmartLabSMS constructor.
      * @param Sender $sender
@@ -38,6 +40,9 @@ class SmartLabSms extends AbstractProvider
         $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
+        $queueName = $this->senderObject->getQueueName();
+        $tries=$this->senderObject->getTries();
+        $backoff=$this->senderObject->getBackoff();
 
         $query = [
             'user' => $config['user'],
@@ -47,7 +52,7 @@ class SmartLabSms extends AbstractProvider
             'smstext' => $text,
         ];
 
-        $requestObject = new Request('https://labapi.smartlabsms.com/smsapi', $query, $queue);
+        $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
         $response = $requestObject->get();
         if ($queue) {
             return true;
