@@ -11,16 +11,17 @@
 
 namespace Xenon\LaravelBDSms\Provider;
 
+use Xenon\LaravelBDSms\Handler\ParameterException;
 use Xenon\LaravelBDSms\Handler\RenderException;
 use Xenon\LaravelBDSms\Request;
 use Xenon\LaravelBDSms\Sender;
 
-class MDL extends AbstractProvider
+class SmsinBD extends AbstractProvider
 {
-    private string $apiEndpoint  = 'http://premium.mdlsms.com/smsapi';
+    private string $apiEndpoint = 'https://api.smsinbd.com/sms-api/sendsms';
 
     /**
-     * MDL constructor.
+     * DianaHost constructor.
      * @param Sender $sender
      */
     public function __construct(Sender $sender)
@@ -30,11 +31,12 @@ class MDL extends AbstractProvider
 
     /**
      * Send Request To Api and Send Message
+     * @throws RenderException
      */
     public function sendRequest()
     {
-        $text = $this->senderObject->getMessage();
         $number = $this->senderObject->getMobile();
+        $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
         $queueName = $this->senderObject->getQueueName();
@@ -42,15 +44,14 @@ class MDL extends AbstractProvider
         $backoff=$this->senderObject->getBackoff();
 
         $query = [
-            'api_key' => $config['api_key'],
-            'type' => $config['type'],
+            'api_token' => $config['api_token'],
             'senderid' => $config['senderid'],
-            'contacts' => $number,
-            'msg' => $text,
+            'contact_number' => $number,
+            'message' => $text,
         ];
 
         $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
-        $response = $requestObject->get();
+        $response = $requestObject->post();
         if ($queue) {
             return true;
         }
@@ -64,19 +65,17 @@ class MDL extends AbstractProvider
     }
 
     /**
-     * @throws RenderException
+     * @throws ParameterException
      */
     public function errorException()
     {
-        if (!array_key_exists('api_key', $this->senderObject->getConfig())) {
-            throw new RenderException('api_key is absent in configuration');
-        }
-        if (!array_key_exists('type', $this->senderObject->getConfig())) {
-            throw new RenderException('type key is absent in configuration');
-        }
-        if (!array_key_exists('senderid', $this->senderObject->getConfig())) {
-            throw new RenderException('senderid key is absent in configuration');
+        if (!array_key_exists('api_token', $this->senderObject->getConfig())) {
+            throw new ParameterException('api_token is absent in configuration');
         }
 
+        if (!array_key_exists('senderid', $this->senderObject->getConfig())) {
+            throw new ParameterException('senderid key is absent in configuration');
+        }
     }
+
 }
