@@ -17,6 +17,8 @@ use Xenon\LaravelBDSms\Sender;
 
 class Ssl extends AbstractProvider
 {
+    private string $apiEndpoint = 'https://smsplus.sslwireless.com/api/v3/send-sms';
+
     /**
      * Ssl constructor.
      * @param Sender $sender
@@ -36,6 +38,9 @@ class Ssl extends AbstractProvider
         $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
+        $queueName = $this->senderObject->getQueueName();
+        $tries = $this->senderObject->getTries();
+        $backoff = $this->senderObject->getBackoff();
 
         $query = [
             'api_token' => $config['api_token'],
@@ -45,11 +50,7 @@ class Ssl extends AbstractProvider
             'sms' => $text,
         ];
 
-        if (is_array($mobile)) {
-            $requestObject = new Request('https://smsplus.sslwireless.com/api/v3/send-sms/bulk', $query, $queue);
-        } else {
-            $requestObject = new Request('https://smsplus.sslwireless.com/api/v3/send-sms', $query, $queue);
-        }
+        $requestObject = new Request($this->apiEndpoint . (is_array($mobile) ? '/bulk' : ''), $query, $queue, [], $queueName, $tries, $backoff);
         $response = $requestObject->post();
         if ($queue) {
             return true;
