@@ -11,16 +11,21 @@
 
 namespace Xenon\LaravelBDSms\Provider;
 
+use Xenon\LaravelBDSms\Handler\ParameterException;
 use Xenon\LaravelBDSms\Handler\RenderException;
 use Xenon\LaravelBDSms\Request;
 use Xenon\LaravelBDSms\Sender;
 
-class MDL extends AbstractProvider
+/**
+ * TwentyFourBulksSMSBD Class
+ * api endpoint https://www.24bulksmsbd.com/api/smsSendApi
+ */
+class TwentyFourBulkSmsBD extends AbstractProvider
 {
-    private string $apiEndpoint  = 'http://premium.mdlsms.com/smsapi';
+    private string $apiEndpoint = 'https://www.24bulksmsbd.com/api/smsSendApi';
 
     /**
-     * MDL constructor.
+     * TwentyFourBulksSMSBD constructor.
      * @param Sender $sender
      */
     public function __construct(Sender $sender)
@@ -30,27 +35,27 @@ class MDL extends AbstractProvider
 
     /**
      * Send Request To Api and Send Message
+     * @return bool|string
+     * @throws RenderException
      */
     public function sendRequest()
     {
-        $text = $this->senderObject->getMessage();
         $number = $this->senderObject->getMobile();
+        $text = $this->senderObject->getMessage();
         $config = $this->senderObject->getConfig();
         $queue = $this->senderObject->getQueue();
         $queueName = $this->senderObject->getQueueName();
-        $tries=$this->senderObject->getTries();
-        $backoff=$this->senderObject->getBackoff();
-
+        $tries = $this->senderObject->getTries();
+        $backoff = $this->senderObject->getBackoff();
         $query = [
+            'customer_id' => $config['customer_id'],
             'api_key' => $config['api_key'],
-            'type' => $config['type'],
-            'senderid' => $config['senderid'],
-            'contacts' => $number,
-            'msg' => $text,
+            'mobile_no' => (is_array($number)) ? implode(',',$number) : $number,
+            'message' => $text,
         ];
 
-        $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName,$tries,$backoff);
-        $response = $requestObject->get();
+        $requestObject = new Request($this->apiEndpoint, $query, $queue, [], $queueName, $tries, $backoff);
+        $response = $requestObject->post();
         if ($queue) {
             return true;
         }
@@ -64,18 +69,16 @@ class MDL extends AbstractProvider
     }
 
     /**
-     * @throws RenderException
+     * @throws ParameterException
      */
     public function errorException()
     {
+        if (!array_key_exists('customer_id', $this->senderObject->getConfig())) {
+            throw new ParameterException('customer_id key is absent in configuration');
+        }
+
         if (!array_key_exists('api_key', $this->senderObject->getConfig())) {
-            throw new RenderException('api_key is absent in configuration');
-        }
-        if (!array_key_exists('type', $this->senderObject->getConfig())) {
-            throw new RenderException('type key is absent in configuration');
-        }
-        if (!array_key_exists('senderid', $this->senderObject->getConfig())) {
-            throw new RenderException('senderid key is absent in configuration');
+            throw new ParameterException('api_key key is absent in configuration');
         }
 
     }
