@@ -7,6 +7,7 @@ class SMS
     /** @var Sender */
     private $sender;
 
+
     /**
      * @param Sender $sender
      * @version v1.0.32
@@ -49,11 +50,19 @@ class SMS
      * @version v1.0.46-dev
      * @since v1.0.46-dev
      */
-    public function shootWithQueue(string $number, string $text)
+    public function shootWithQueue(string $number, string $text, string $queueName = 'default', int $tries = 3, int $backoff = 60)
     {
         $this->sender->setMobile($number);
         $this->sender->setMessage($text);
         $this->sender->setQueue(true);
+        $this->sender->setQueueName($queueName);
+
+        if (isset($tries)) {
+            $this->sender->setTries($tries);
+        }
+        if (isset($backoff)) {
+            $this->sender->setBackoff($backoff);
+        }
         return $this->sender->send();
     }
 }
